Modify X-Rays to be reusable in other apps.
