+++ xray/src/lib.rs
@@ -1,6 +1,15 @@
 extern crate cgmath;
+extern crate collision;
+#[macro_use]
+extern crate iron;
 extern crate protobuf;
 extern crate quadtree;
+extern crate router;
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+extern crate urlencoded;
 
 use cgmath::Point2;
 use quadtree::{NodeId, Rect};
@@ -56,5 +65,7 @@ impl Meta {
         }
     }
 }
+
+pub mod backend;
 
 include!(concat!(env!("OUT_DIR"), "/proto.rs"));
